filter by multples of 3,5, and none
