+++ scripts/fish/FishDataProvider.js
@@ -2,7 +2,7 @@ const fishCollection = [
     {
         name: "Bart",
         species: "Clownfish",
-        length: 4,
+        length: 5,
         location: "Great Barrier Reef",
         food: "small invertebrates and algae",
         image: "clownfish.jpg"
@@ -36,3 +36,40 @@ const fishCollection = [
 export const useFish = () =>{
     return fishCollection.slice();
 }
+
+//chapter 13 Big Fish, Little Fish&& this code filters fish by their length 
+// if fish length is a multiple of three, then place in holyFish array
+export const mostHolyFish = () => {
+    const holyFish = []
+
+    for (const fish of fishCollection) {
+        if(fish.length % 3 === 0){
+            holyFish.push(fish)
+        }   
+    }
+    return holyFish
+}
+
+//if fish length is a multiple of five, then place in soldier array
+export const soldierFish = () => {
+    const soldiers = []
+    for (const fish of fishCollection){
+        if (fish.length % 5 === 0 ){
+            soldiers.push(fish)
+        }
+    }  
+    return soldiers
+}
+
+//If fish length is not a multiple of three or fish, then place in regularFish array
+export const nonHolyFish = () => {
+    const regularFish = []
+    for(const fish of fishCollection){
+        if (fish.length % 3 != 0 && fish.length % 5 != 0){
+            regularFish.push(fish)
+        }
+    }
+    return regularFish
+}
+
+
